requirements.txt
spoon-ai-sdk>=0.1.0
spoon-toolkits>=0.1.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
pydantic>=2.0.0
flask>=3.0.0
flask-cors>=4.0.0
